Se solucinan errores

+++ src/main/java/com/examen/examen_final/models/Pago.java
@@ -17,5 +17,5 @@ public class Pago {
     private int tipo_pago_id;
     private String tarjeta_tipo;
     private int cuotas;
-    private BigDecimal valor;
+    private long valor;
 }
